Same change as before in other functions
Group.GetMinGroup, Group.GetMaxGroup, Group.GetSum, Group.GetClass, Group.GetClasCount

+++ libgroup.cpp
@@ -574,7 +574,7 @@ LUA_FUNCTION(GetMinGroup) {
 	if(pgroup->container.size() == 0)
 		return 0;
 	group* newgroup = pduel->new_group();
-	int32_t min, op;
+	int64_t min, op;
 	int32_t extraargs = lua_gettop(L) - 2;
 	auto cit = pgroup->container.begin();
 	min = pduel->lua->get_operation_value(*cit, findex, extraargs);
@@ -602,7 +602,7 @@ LUA_FUNCTION(GetMaxGroup) {
 	if(pgroup->container.size() == 0)
 		return 0;
 	group* newgroup = pduel->new_group();
-	int32_t max, op;
+	int64_t max, op;
 	int32_t extraargs = lua_gettop(L) - 2;
 	auto cit = pgroup->container.begin();
 	max = pduel->lua->get_operation_value(*cit, findex, extraargs);
@@ -628,7 +628,7 @@ LUA_FUNCTION(GetSum) {
 	auto pgroup = lua_get<group*, true>(L, 1);
 	const auto pduel = lua_get<duel*>(L);
 	int32_t extraargs = lua_gettop(L) - 2;
-	int32_t sum = 0;
+	int64_t sum = 0;
 	for(auto& pcard : pgroup->container) {
 		sum += pduel->lua->get_operation_value(pcard, findex, extraargs);
 	}
@@ -667,7 +667,7 @@ LUA_FUNCTION(GetClass) {
 	auto pgroup = lua_get<group*, true>(L, 1);
 	const auto pduel = lua_get<duel*>(L);
 	int32_t extraargs = lua_gettop(L) - 2;
-	std::set<uint32_t> er;
+	std::set<int64_t> er;
 	for(auto& pcard : pgroup->container) {
 		er.insert(pduel->lua->get_operation_value(pcard, findex, extraargs));
 	}
@@ -686,7 +686,7 @@ LUA_FUNCTION(GetClassCount) {
 	auto pgroup = lua_get<group*, true>(L, 1);
 	const auto pduel = lua_get<duel*>(L);
 	int32_t extraargs = lua_gettop(L) - 2;
-	std::set<uint32_t> er;
+	std::set<int64_t> er;
 	for(auto& pcard : pgroup->container) {
 		er.insert(pduel->lua->get_operation_value(pcard, findex, extraargs));
 	}
